// src/maracas-schema.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface components {
  schemas: {
    BreakingChangeDto: {
      declaration: string;
      change: string;
      path: string;
      startLine: number;
      endLine: number;
      fileUrl: string;
      diffUrl: string;
    };
    BrokenUseDto: {
      elem: string;
      used: string;
      src: string;
      apiUse: string;
      path: string;
      startLine: number;
      endLine: number;
      url: string;
    };
    ClientReport: {
      url: string;
      brokenUses: (components['schemas']['BrokenUseDto'])[];
      error: string | undefined | null;
    };
    DeltaDto: {
      jarV1: string;
      jarV2: string;
      breakingChanges: (components['schemas']['BreakingChangeDto'])[];
    };
    MaracasReport: {
      reports: (components['schemas']['PackageReport'])[];
    };
    PackageReport: {
      id: string;
      error: string | undefined | null;
      delta: components['schemas']['DeltaDto'] | undefined | null;
      clientReports: (components['schemas']['ClientReport'])[];
    };
    PullRequestDto: {
      owner: string;
      name: string;
      number: number;
      baseBranch: string;
      headBranch: string;
      baseSha: string;
      headSha: string;
    };
    PullRequestResponse: {
      message: string;
      date: string;
      pr: components['schemas']['PullRequestDto'] | undefined | null;
      report: components['schemas']['MaracasReport'] | undefined | null;
    };
    Client: {
      pkg: components['schemas']['Package'];
      owner: string;
      name: string;
      stars: number;
      forks: number;
    };
    ClientsResponse: {
      message: string;
      owner: string;
      name: string;
      packages: (components['schemas']['Package'])[];
      clients: (components['schemas']['Client'])[];
    };
    Package: {
      name: string;
      url: string;
    };
  };
}